requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.2
aiosignal==1.3.1
attrs==24.2.0
certifi==2024.7.4        
charset-normalizer==3.3.2
discord.py==2.4.0        
frozenlist==1.4.1        
idna==3.7
multidict==6.0.5
python-dotenv==1.0.1     
requests==2.32.3
urllib3==2.2.2
yarl==1.9.4